requirements.txt
asgiref==3.7.2
boto3==1.14.32
botocore==1.17.32
channels==4.0.0
daphne
Django==5.0.1
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-imagekit
django-storages
django-taggit
docutils
gunicorn
jmespath
pilkit
pillow
python-dotenv
pytz
s3transfer
six
sqlparse
urllib3
django-cloudinary-storage
dj-database-url
psycopg2-binary